requirements.txt
absl-py==2.0.0
amqp==5.2.0
asgiref==3.7.2
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
billiard==4.2.0
biopython==1.81
biotite==0.38.0
blinker==1.7.0
cachetools==5.3.1
celery==5.3.6
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Django==4.2.4
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-redis==5.4.0
django-taggit==5.0.1
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
Flask==3.0.2
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
kombu==5.3.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
MouseInfo==0.1.3
msgpack==1.0.7
networkx==3.1
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
patsy==0.5.3
paynow==1.0.6
phonenumbers==8.13.31
Pillow==10.0.0
prompt-toolkit==3.0.43
protobuf==4.24.4
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
PyBrain==0.3
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.2.0
pytz==2023.3.post1
pywhatkit==5.4
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
zipp==3.17.0
